+++ src/dlc/UI/Shared/CivVAccess_LeaderDescription.lua
@@ -68,7 +68,7 @@ function LeaderDescription.bindF2(handler, getPlayerIdFn)
     }
     BaseMenuHelp.addScreenKey(handler, {
         keyLabel = "TXT_KEY_CIVVACCESS_HELP_KEY_F2",
-        description = "TXT_KEY_CIVVACCESS_HELP_DESC_LEADER_DESC",
+        description = "TXT_KEY_CIVVACCESS_HELP_DESC_IMAGE_DESC",
     })
 end